Move generated types from JSON schema to `generated/app-webhooks-types` folder.
