+++ .changeset/khaki-pans-own.md
@@ -1,0 +1,5 @@
+---
+"saleor-app-payment-stripe": patch
+---
+
+Move generated types from JSON schema to `generated/app-webhooks-types` folder.
