+++ apps/stripe/src/app/api/webhooks/saleor/payment-gateway-initialize-session/use-case-response.ts
@@ -1,7 +1,7 @@
 import { z } from "zod";
 
 import { SuccessWebhookResponse } from "@/app/api/webhooks/saleor/saleor-webhook-responses";
-import { PaymentGatewayInitializeSession } from "@/generated/json-schema/payment-gateway-initialize-session";
+import { PaymentGatewayInitializeSession } from "@/generated/app-webhooks-types/payment-gateway-initialize-session";
 import { AppContext } from "@/lib/app-context";
 import { StripePublishableKey } from "@/modules/stripe/stripe-publishable-key";
