+++ apps/stripe/src/modules/transaction-result/refund-result.ts
@@ -1,4 +1,4 @@
-import { Actions } from "@/generated/json-schema/transaction-refund-requested";
+import { Actions } from "@/generated/app-webhooks-types/transaction-refund-requested";
 
 export class RefundSuccessResult {
   readonly result = "REFUND_SUCCESS" as const;
