+++ apps/stripe/src/modules/transaction-result/success-result.ts
@@ -1,4 +1,4 @@
-import { Actions } from "@/generated/json-schema/transaction-charge-requested";
+import { Actions } from "@/generated/app-webhooks-types/transaction-charge-requested";
 
 export class ChargeSuccessResult {
   readonly result = "CHARGE_SUCCESS" as const;
